Improvements to bootstrapping
Have added some explanation of how to configure a new Ubuntu box from
scratch so that it will work with vagrant. Also a fair bit of general
configuration ‘stuff’.  Normally, I’d expect to be the only one to run
this, but I figure that I’ll have to do it again so why not document it
and make it as easy as possible to provision new versions.